+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Remove empty declaration from stylesheet
     - Review instanciation vs. configuration parameters, introducing setupEditor() method (todo #14)
     - Have a warning when tabular name doesn't have been registered
+    - Update transformation to incoming pwix:roles v1.10
 
 ### 2.4.0